requirements.txt
cfgv==3.4.0
distlib==0.3.9
filelock==3.16.1
greenlet==3.1.1
identify==2.6.1
mypy==1.13.0
mypy-extensions==1.0.0
nodeenv==1.9.1
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
PyYAML==6.0.2
ruff==0.7.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
virtualenv==20.27.1
